Only warn once per type variable
To accomplish this, we split the type-indexed `TyVarNames` into `Binder(s)`
and `UnquantifiedTyVar(s)`. The advantage is that we can now use a `NameSet`
for `Binders`, as we don't care for their ordering, and a left-biased
`Data.Set` for `UnquantifiedTyVars`, so we only retain the first occurrence of
an unquantified type variable in a signature.